chore: move types files
